requirements.txt
python-docx
pdfplumber
docxtpl
google-generativeai
python-dotenv
